requirements.txt
numpy
tornado
pyserial
wxPython
matplotlib
PySimpleGUI
